Fix signedness & usage warnings

+++ tools/unix_shims_randombytes.c
@@ -5,6 +5,6 @@ void randombytes_buf(void * const buf, const size_t size) {
     FILE *urandom = fopen("/dev/urandom", "r");
     assert(urandom != NULL);
     int count = fread(buf, 1, size, urandom);
-    assert(count == size);
+    assert(count == (int)size);
     fclose(urandom);
 }
